+++ src/ansys/fluent/core/_grpc_services/health_check_service.py
@@ -1,0 +1,144 @@
+# Copyright (C) 2021 - 2026 ANSYS, Inc. and/or its affiliates.
+# SPDX-License-Identifier: MIT
+#
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+"""Wrapper over the health check gRPC service of Fluent (v1 proto API)."""
+
+from enum import Enum
+import logging
+import sys
+
+import grpc
+
+from ansys.api.fluent.v1 import health_pb2, health_pb2_grpc
+from ansys.fluent.core.module_config import config
+from ansys.fluent.core.services._protocols import ServiceProtocol
+from ansys.fluent.core.services.interceptors import (
+    BatchInterceptor,
+    ErrorStateInterceptor,
+    GrpcErrorInterceptor,
+    TracingInterceptor,
+)
+
+logger: logging.Logger = logging.getLogger("pyfluent.general")
+
+
+class HealthCheckService(ServiceProtocol):
+    """Class wrapping the health check gRPC service of Fluent.
+
+    Methods
+    -------
+    check_health()
+        Check the health of the Fluent connection.
+    """
+
+    class Status(Enum):
+        """Health check status."""
+
+        SERVING: int = 1
+        NOT_SERVING: int = 2
+
+    def __init__(
+        self, channel: grpc.Channel, metadata: list[tuple[str, str]], fluent_error_state
+    ) -> None:
+        """__init__ method of HealthCheckService class."""
+        intercept_channel = grpc.intercept_channel(
+            channel,
+            GrpcErrorInterceptor(),
+            ErrorStateInterceptor(fluent_error_state),
+            TracingInterceptor(),
+            BatchInterceptor(),
+        )
+        self._stub = health_pb2_grpc.HealthStub(intercept_channel)
+        self._metadata = metadata
+        self._channel = channel
+
+    def _create_health_check_request(self):
+        """Create a health-check request. Override in subclasses for different proto modules."""
+        return health_pb2.HealthCheckRequest()
+
+    def check_health(self) -> Status:
+        """Check the health of the Fluent connection.
+
+        Returns
+        -------
+        Status
+        """
+        request = self._create_health_check_request()
+        response = self._stub.Check(
+            request,
+            metadata=self._metadata,
+            timeout=config.check_health_timeout,
+        )
+        return self.Status(response.status)
+
+    def wait_for_server(self, timeout: int) -> None:
+        """Keeps a watch on the health of the Fluent connection.
+
+        Response changes only when the service's serving status changes.
+
+        Parameters
+        ----------
+        timeout : int
+            timeout in seconds
+
+        Raises
+        ------
+        TimeoutError
+            If the connection to the Fluent server could not be established within the timeout.
+        """
+        request = self._create_health_check_request()
+        responses = self._stub.Watch(request, metadata=self._metadata, timeout=timeout)
+
+        while True:
+            try:
+                response = next(responses)
+                if response.status == 1:
+                    responses.cancel()
+            except StopIteration:
+                break
+            except grpc.RpcError as e:
+                if e.code() == grpc.StatusCode.CANCELLED:
+                    break
+                if e.code() == grpc.StatusCode.DEADLINE_EXCEEDED:
+                    raise TimeoutError(
+                        f"The connection to the Fluent server could not be established within the configurable {timeout} second time limit."
+                    )
+                raise
+
+    def status(self) -> Status:
+        """Check health of Fluent connection."""
+        if self._channel:
+            try:
+                return self.check_health()
+            except Exception:
+                ex_type, ex_value, _ = sys.exc_info()
+                logger.info(
+                    f"HealthCheckService.status() caught {ex_type.__name__}: {ex_value}"
+                )
+                return self.Status.NOT_SERVING
+        else:
+            return self.Status.NOT_SERVING
+
+    @property
+    def is_serving(self) -> bool:
+        """Checks whether Fluent is serving."""
+        return self.status() == self.Status.SERVING
