+++ src/ansys/fluent/core/services/__init__.py
@@ -58,18 +58,8 @@ from ansys.fluent.core.services.field_data_v1 import (
     LiveFieldData,
     _FieldInfo,
 )
-from ansys.fluent.core.services.health_check import (
-    HealthCheckService as HealthCheckServiceV0,
-)
-from ansys.fluent.core.services.health_check_v1 import HealthCheckService
 from ansys.fluent.core.services.monitor import MonitorsService as MonitorsServiceV0
 from ansys.fluent.core.services.monitor_v1 import MonitorsService as MonitorsService
-from ansys.fluent.core.services.reduction import Reduction as ReductionV0
-from ansys.fluent.core.services.reduction_v1 import Reduction
-from ansys.fluent.core.services.scheme_eval import (
-    SchemeEvalService as SchemeEvalServiceV0,
-)
-from ansys.fluent.core.services.scheme_eval_v1 import SchemeEvalService
 from ansys.fluent.core.services.settings import SettingsService as SettingsServiceV0
 from ansys.fluent.core.services.settings_v1 import SettingsService
 from ansys.fluent.core.services.solution_variables import (
@@ -137,17 +127,14 @@ __all__ = (
 
 _service_cls_by_name_v0 = {
     "application_runtime": ApplicationRuntimeServiceV0,
-    "health_check": HealthCheckServiceV0,
     "datamodel": DatamodelService_SE_V0,
     "tui": DatamodelService_TUI_V0,
     "settings": SettingsServiceV0,
-    "scheme_interpreter": SchemeEvalServiceV0,
     "events": EventsServiceV0,
     "field_data": LiveFieldDataV0,
     "field_data_old": DeprecatedFieldData,
     "field_info": _FieldInfoV0,
     "monitors": MonitorsServiceV0,
-    "reduction": ReductionV0,
     "svar": SolutionVariableServiceV0,
     "svar_data": SolutionVariableDataV0,
     "transcript": TranscriptServiceV0,
@@ -157,17 +144,14 @@ _service_cls_by_name_v0 = {
 
 _service_cls_by_name = {
     "application_runtime": ApplicationRuntimeService,
-    "health_check": HealthCheckService,
     "datamodel": DatamodelService_SE,
     "tui": DatamodelService_TUI,
     "settings": SettingsService,
-    "scheme_interpreter": SchemeEvalService,
     "events": EventsService,
     "field_data": LiveFieldData,
     "field_data_old": DeprecatedFieldData,
     "field_info": _FieldInfo,
     "monitors": MonitorsService,
-    "reduction": Reduction,
     "svar": SolutionVariableService,
     "svar_data": SolutionVariableData,
     "transcript": TranscriptService,
