+++ src/ansys/fluent/core/session_solver.py
@@ -37,10 +37,7 @@ from ansys.fluent.core.pyfluent_warnings import PyFluentDeprecationWarning
 from ansys.fluent.core.services import MonitorsServiceV0, service_creator
 from ansys.fluent.core.services.field_data import ZoneInfo, ZoneType
 from ansys.fluent.core.services.monitor_v1 import MonitorsService
-from ansys.fluent.core.services.reduction import Reduction as ReductionV0
-from ansys.fluent.core.services.reduction import ReductionService as ReductionServiceV0
-from ansys.fluent.core.services.reduction_v1 import Reduction, ReductionService
-from ansys.fluent.core.services.scheme_eval_v1 import SchemeEval
+from ansys.fluent.core.services.scheme_interpreter import SchemeInterpreter
 from ansys.fluent.core.services.solution_variables import (
     SolutionVariableData as SolutionVariableDataV0,
 )
@@ -120,7 +117,7 @@ class Solver(BaseSession, settings_root.root if TYPE_CHECKING else object):
     def __init__(
         self,
         fluent_connection,
-        scheme_eval: SchemeEval,
+        scheme_eval: SchemeInterpreter,
         file_transfer_service: Any | None = None,
         start_transcript: bool = True,
         launcher_args: dict[str, Any] | None = None,
@@ -131,8 +128,8 @@ class Solver(BaseSession, settings_root.root if TYPE_CHECKING else object):
         ----------
         fluent_connection (:ref:`ref_fluent_connection`):
             Encapsulates a Fluent connection.
-        scheme_eval: SchemeEval
-            Instance of ``SchemeEval`` to execute Fluent's scheme code on.
+        scheme_eval: SchemeInterpreter
+            Instance of ``SchemeInterpreter`` to execute Fluent's scheme code on.
         file_transfer_service : Optional
             Service for uploading and downloading files.
         start_transcript : bool, optional
@@ -161,7 +158,7 @@ class Solver(BaseSession, settings_root.root if TYPE_CHECKING else object):
     def _build_from_fluent_connection(
         self,
         fluent_connection: "FluentConnection",
-        scheme_eval: SchemeEval,
+        scheme_eval: SchemeInterpreter,
         file_transfer_service: Any | None = None,
         launcher_args: dict[str, Any] | None = None,
     ):
@@ -176,19 +173,13 @@ class Solver(BaseSession, settings_root.root if TYPE_CHECKING else object):
         self._solution_variable_service = service_creator(
             "svar", supports_v1=fluent_connection._server_supports_v1
         ).create(fluent_connection._channel, fluent_connection._metadata)
+        self.fields.reduction = fluent_connection._service_factory.reduction
+        self.fields.reduction.set_context(self)
         if fluent_connection._server_supports_v1:
-            self._reduction_service = fluent_connection.create_grpc_service(
-                ReductionService, self._error_state
-            )
-            self.fields.reduction = Reduction(self._reduction_service, self)
             self.fields.solution_variable_info = SolutionVariableInfo(
                 self._solution_variable_service
             )
         else:
-            self._reduction_service = fluent_connection.create_grpc_service(
-                ReductionServiceV0, self._error_state
-            )
-            self.fields.reduction = ReductionV0(self._reduction_service, self)
             self.fields.solution_variable_info = SolutionVariableInfoV0(
                 self._solution_variable_service
             )
