+++ tests/test_scheme_eval_231.py
@@ -26,7 +26,7 @@ from google.protobuf.json_format import MessageToDict, ParseDict
 import pytest
 
 from ansys.api.fluent.v0.scheme_pointer_pb2 import SchemePointer
-from ansys.fluent.core.services.scheme_eval import (
+from ansys.fluent.core._grpc_services.scheme_interpreter import (
     Symbol,
     _convert_py_value_to_scheme_pointer,
     _convert_scheme_pointer_to_py_value,
